[clean] PR corrections
